Added a function that prints an integer.

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,7 +7,7 @@
  */
 void print_number(int n)
 {
-	signed int num;
+	short int num;
 
 	if (n < 0)
 	{
